+++ np.Templating/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 See Plugin [Documentation](https://noteplan.co/templates/docs) for details on available commands and use case.
 
+## [2.0.19] 2025-08-06 @dwertheimer
+- Fix calendar notes not working right after title searching changes
+- Fix edge case where a template had -- frontmatter but no tags and was taking the fast path wrongly
+
 ## [2.0.18] 2025-08-06 @dwertheimer
 - add pluginID for iphone settings editing